refactor: ViewModelFactory를 di 패키지로 이동
