+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -10,7 +10,7 @@ import androidx.core.view.WindowInsetsCompat
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
 import woowacourse.shopping.di.AppContainer
-import woowacourse.shopping.ui.ViewModelFactory
+import woowacourse.shopping.di.ViewModelFactory
 
 class CartActivity : AppCompatActivity() {
     private val binding by lazy { ActivityCartBinding.inflate(layoutInflater) }
